Cascading tables are working. Added more examples. Cascading recipe search starting to look good.
