+++ Tests/Dev_Tests.cs
@@ -13,6 +13,8 @@ namespace GilGoblin.Tests
         // Item ID for Mithril Ore: 5114
         // Item ID for Copper Ore: 5106
         // Item ID for Copper Ingot: 5057
+        // Item ID for Chondrite Magitek Sword: 34680
+        // Item ID for Bomb Frypan: 2499
 
         public const string world_name = "Brynhildr";
         public const int world_id = 34;
@@ -62,33 +64,15 @@ namespace GilGoblin.Tests
             try
             {
                 Log.Logger = new LoggerConfiguration()
-                    .MinimumLevel.Debug()
-                    // PROD: reduce verbosity
+                    .MinimumLevel.Debug() // PROD: reduce verbosity
                     .WriteTo.Console()  
                     .WriteTo.File("logs/test.txt")
                     .CreateLogger();
 
                 Log.Debug("Application started.");
 
-                List<int> fetchIDList = new List<int> { item_id, 5114, 5106 };
-                //List<ItemInfo> infoList = new List<ItemInfo>();
-                //List<MarketDataDB> marketDataList 
-                //    = MarketData.GetMarketDataBulk(fetchIDList, world_id, true);
-                //foreach (MarketDataDB dataDb in marketDataList)
-                //{
-                //    infoList.Add(ItemInfoDB.GetItemInfo(item_id));
-                //}
-
-                //if (infoList.Count == 0 || marketDataList.Count == 0)
-                //{
-                //    throw new Exception("Market data or item info not found");
-                //}
-
-                //string success_message = "failure.";
-                //int db_success = await DatabaseAccess.SaveMarketDataBulk(marketDataList);
-                //if (db_success > 0) { success_message = "sucess."; }
-                //Log.Information("Database save was a " + success_message);
-                //Log.Information(db_success + " entries saved to the database.");
+                List<int> fetchIDList 
+                    = new List<int> { 2499, 34680, 5057, 5114, 5106 };
 
                 foreach (int id in fetchIDList) 
                 {
